+++ Node/ecommerce-api/src/services/category.service.ts
@@ -1,12 +1,16 @@
 import { NotFoundError } from "../errors/not-found.error.js";
+import { ValidationError } from "../errors/validation-error.js";
 import { Category } from "../models/category.model.js";
 import { CategoryRepository } from "../repositories/category.repository.js";
+import { ProductRepository } from "../repositories/product.repository.js";
 
 export class CategoryService {
   private categoryRepository: CategoryRepository;
+  private productRepository: ProductRepository;
 
   constructor() {
     this.categoryRepository = new CategoryRepository();
+    this.productRepository = new ProductRepository();
   }
 
   async getAll(): Promise<Category[]> {
@@ -36,6 +40,9 @@ export class CategoryService {
   }
 
   async delete(id: string) {
+    if(await this.productRepository.getCountByCategoria(id) > 0) {
+      throw new ValidationError("Não é possível excluir uma categoria com produtos relacionados!");
+    }
     await this.categoryRepository.delete(id);
   }
 }
